Rework locale specific pluralization rules
I regenerated pluralization rules according to internal sources that are
used in our non-JavaScript environments. This is needed to ensure
consistency between translations regardless of platform. The plural rule
functions were automatically generated from a Java source, so they are
somewhat more verbose and less hand-factored.

Names of plural types are purely mnemonic and do not imply a specific
relationship between two languages, only that their pluralType functions
would be the same.

Polyglot supports locales that we don't use internally that were
contributed by the community. I've ported those over to maintain
compatibility. There were several places where exact locales were
specified. I changed those to be based on language-only when doing so
would not change the result. Behavior for pt-BR is preserved, as it does
have a different pluralization rule than pt.

There were a couple areas of disagreement between sources. In these
cases, I went with what we were using internally, which tends to be close to
https://developer.mozilla.org/en-US/docs/Mozilla/Localization/Localization_and_Plurals

The specific changes in this PR are

1. tr from germanLike to chineseLike
2. lt rule changed

Additionally, per our records, fa would be chineseLike instead of
germanLike, but as we do not yet support this language, I have left it
as is.